[Feat] #9 BaseTableViewCell 정의, UITableViewCell 객체 상속 및 구현
