+++ MeaningOut/Sources/Presentation/Common/Base/BaseTableViewCell.swift
@@ -1,0 +1,25 @@
+//
+//  BaseTableViewCell.swift
+//  MeaningOut
+//
+//  Created by gnksbm on 6/27/24.
+//
+
+import UIKit
+
+class BaseTableViewCell: UITableViewCell {
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        configureUI()
+        configureLayout()
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configureUI() { }
+    func configureLayout() { }
+}
+
