+++ MeaningOut/Sources/Presentation/MainTab/Setting/Cell/SettingTableViewCountCell.swift
@@ -14,7 +14,7 @@ protocol CountCellData {
     var itemName: String { get }
 }
 
-final class SettingTableViewCountCell: UITableViewCell {
+final class SettingTableViewCountCell: BaseTableViewCell {
     private let descriptionLabel = UILabel().build { builder in
         builder.font(DesignConstant.Font.large.with(weight: .regular))
     }
@@ -28,16 +28,6 @@ final class SettingTableViewCountCell: UITableViewCell {
             .required,
             for: .horizontal
         )
-    }
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        configureLayout()
-        selectionStyle = .none
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
     }
     
     override func prepareForReuse() {
@@ -79,7 +69,11 @@ final class SettingTableViewCountCell: UITableViewCell {
         }
     }
     
-    private func configureLayout() {
+    override func configureUI() {
+        selectionStyle = .none
+    }
+    
+    override func configureLayout() {
         [
             descriptionLabel,
             iconImageView,
